+++ clients/derivatives-trading-portfolio-margin/src/rest-api/types/new-margin-order-response.ts
@@ -98,10 +98,10 @@ export interface NewMarginOrderResponse {
     side?: string;
     /**
      *
-     * @type {number}
+     * @type {string}
      * @memberof NewMarginOrderResponse
      */
-    marginBuyBorrowAmount?: number;
+    marginBuyBorrowAmount?: string;
     /**
      *
      * @type {string}