+++ clients/derivatives-trading-portfolio-margin/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 6.0.0 - 2025-09-17
+
+### Changed (2)
+
+#### REST API
+
+- Modified response for `marginMaxBorrow()` (`GET /papi/v1/margin/maxBorrowable`):
+  - `amount`: type `number` → `string`
+  - `borrowLimit`: type `integer` → `string`
+
+- Modified response for `newMarginOrder()` (`POST /papi/v1/margin/order`):
+  - `marginBuyBorrowAmount`: type `integer` → `string`
+
 ## 5.0.5 - 2025-09-12
 
 ### Changed (1)